+++ api/func/auth.py
@@ -1,11 +1,11 @@
 from flask import redirect, request, jsonify
 from urllib.parse import urlencode
-from __main__ import client_id, app
+from __main__ import client_id
 from func.code import codeChallenge, codeVerifier
 import requests
 
 redirect_uri = "http://localhost:3000"
-scope = "user-read-private user-read-email playlist-modify-public playlist-modify-private"
+scope = "user-read-private user-read-email user-read-playback-state user-read-currently-playing"
 auth_url = "https://accounts.spotify.com/authorize"
 
 def get_auth_url():
@@ -25,7 +25,7 @@ def auth_route():
     auth_url = get_auth_url()
     return redirect(auth_url)
 
-def get_token(code):
+def get_code_token(code):
     token_url = "https://accounts.spotify.com/api/token"
     
     payload = {
@@ -51,19 +51,28 @@ def get_token(code):
         }
         
 
-def callback_route():
-    code = request.args.get("code")
-    if code:
-        token_response = get_token(code)
-        return jsonify(token_response)
-    return jsonify({
-        "error": "No code provided"
-    })
+# def callback_route():
+#     code = request.args.get("code")
+#     if code:
+#         token_response = get_code_token(code)
+#         return jsonify(token_response)
+#     return jsonify({
+#         "error": "No code provided"
+#     })
     
-@app.route("/api/auth")
-def auth():
-    return auth_route()
+    
+# @app.route("/api/auth")
+# def auth():
+#     return auth_route()
 
-@app.route("/api/callback")
-def callback():
-    return callback_route()
+# @app.route("/api/callback")
+# def callback():
+#     return callback_route()
+
+# @app.route("/api/token")
+# def get_token_route():
+#     data = {
+#         "token": get_token()
+#     }
+    
+#     return jsonify(data)