+++ api/func/header.py
@@ -1,5 +1,5 @@
 # This function retrieves the authorization header for the Spotify API.
 def get_auth_header(token):
     return {
-        "Authorization": "Bearer " + token
+        "Authorization": f"Bearer {token}"
     }