+++ app/layout.tsx
@@ -1,5 +1,13 @@
 import "./globals.css";
 import type { Metadata } from "next";
+import { Montserrat } from "next/font/google";
+
+
+const montserrat = Montserrat({
+    subsets: ["latin"],
+    weight: ["300", "400", "500", "600", "700"],
+    variable: "--font-mont",
+})
 
 export const metadata: Metadata = {
     title: "Spotify API Starter",
@@ -15,7 +23,9 @@ export default async function RootLayout({
             <head>
                 
             </head>
-            <body>{children}</body>
+            <body className={montserrat.variable}>
+                {children}
+            </body>
         </html>
     )
 }